move jsonwebtoken package
