+++ packages/next/server/api-utils/index.ts
@@ -1,9 +1,7 @@
-import type { IncomingMessage, ServerResponse } from 'http'
-import type { PreviewData } from 'next/types'
-import type { BaseNextRequest, BaseNextResponse } from '../base-http'
+import type { IncomingMessage } from 'http'
+import type { BaseNextRequest } from '../base-http'
 
 import { NextApiRequest, NextApiResponse } from '../../shared/lib/utils'
-import { decryptWithSecret } from '../crypto-utils'
 
 export type NextApiRequestCookies = { [key: string]: string }
 export type NextApiRequestQuery = { [key: string]: string | string[] }
@@ -86,82 +84,6 @@ export const COOKIE_NAME_PRERENDER_DATA = `__next_preview_data`
 
 export const SYMBOL_PREVIEW_DATA = Symbol(COOKIE_NAME_PRERENDER_DATA)
 export const SYMBOL_CLEARED_COOKIES = Symbol(COOKIE_NAME_PRERENDER_BYPASS)
-
-export function tryGetPreviewData(
-  req: IncomingMessage | BaseNextRequest,
-  res: ServerResponse | BaseNextResponse,
-  options: __ApiPreviewProps
-): PreviewData {
-  // Read cached preview data if present
-  if (SYMBOL_PREVIEW_DATA in req) {
-    return (req as any)[SYMBOL_PREVIEW_DATA] as any
-  }
-
-  const getCookies = getCookieParser(req.headers)
-  let cookies: NextApiRequestCookies
-  try {
-    cookies = getCookies()
-  } catch {
-    // TODO: warn
-    return false
-  }
-
-  const hasBypass = COOKIE_NAME_PRERENDER_BYPASS in cookies
-  const hasData = COOKIE_NAME_PRERENDER_DATA in cookies
-
-  // Case: neither cookie is set.
-  if (!(hasBypass || hasData)) {
-    return false
-  }
-
-  // Case: one cookie is set, but not the other.
-  if (hasBypass !== hasData) {
-    clearPreviewData(res as NextApiResponse)
-    return false
-  }
-
-  // Case: preview session is for an old build.
-  if (cookies[COOKIE_NAME_PRERENDER_BYPASS] !== options.previewModeId) {
-    clearPreviewData(res as NextApiResponse)
-    return false
-  }
-
-  const tokenPreviewData = cookies[COOKIE_NAME_PRERENDER_DATA]
-
-  const jsonwebtoken =
-    require('next/dist/compiled/jsonwebtoken') as typeof import('jsonwebtoken')
-  let encryptedPreviewData: {
-    data: string
-  }
-  try {
-    encryptedPreviewData = jsonwebtoken.verify(
-      tokenPreviewData,
-      options.previewModeSigningKey
-    ) as typeof encryptedPreviewData
-  } catch {
-    // TODO: warn
-    clearPreviewData(res as NextApiResponse)
-    return false
-  }
-
-  const decryptedPreviewData = decryptWithSecret(
-    Buffer.from(options.previewModeEncryptionKey),
-    encryptedPreviewData.data
-  )
-
-  try {
-    // TODO: strict runtime type checking
-    const data = JSON.parse(decryptedPreviewData)
-    // Cache lookup
-    Object.defineProperty(req, SYMBOL_PREVIEW_DATA, {
-      value: data,
-      enumerable: false,
-    })
-    return data
-  } catch {
-    return false
-  }
-}
 
 export function clearPreviewData<T>(
   res: NextApiResponse<T>
